pyproject.toml
[project]
name = "xlsform-ai-cli"
version = "0.6.0"
description = "AI-powered XLSForm creation tool with Claude Code integration. An open source project by ARCED International."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "ARCED International", email = "info@arced-international.com" }
]
keywords = ["xlsform", "odk", "survey", "cli", "ai", "humanitarian", "m&e"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
    "questionary>=2.0.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "pyxform>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
ai = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
]
ai-gpu = [
    "sentence-transformers>=2.2.0",
    "faiss-gpu>=1.7.4",
]
translate = [
    "deep-translator>=1.11.4",
]

[project.scripts]
xlsform-ai = "xlsform_ai.cli:app"

[project.urls]
Homepage = "https://github.com/ARCED-International/xlsform-ai"
Repository = "https://github.com/ARCED-International/xlsform-ai"
Documentation = "https://github.com/ARCED-International/xlsform-ai/wiki"
Issues = "https://github.com/ARCED-International/xlsform-ai/issues"
Organization = "https://arced-international.com"

[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
xlsform_ai = ["templates/*.xlsx", "templates/*.md", "templates/*.json", "templates/*.py", "templates/*.html",
              "templates/**/*.xlsx", "templates/**/*.md", "templates/**/*.json", "templates/**/*.py", "templates/**/*.html"]

[tool.setuptools_scm]
allow_dirty = true
